Separate method and property info into separate files
